// src/contants/enum.ts
export enum MessageTypeEnum {
  OPTION = 'option',
  TEXT = 'text',
  CONFIRM = 'confirm',
  IMAGE = 'image',
  CHOOSE_IMG = 'choose-img',
  TYPING = 'typing',
  USER = 'user',
  COMPLETED = 'completed',
}

export enum HistoryChatEnum {
  MESSAGES = 'messages',
  PROMPTS = 'prompts',
  LEVEL = 'level',
}